feat: add product constructor

+++ produto/src/Main.java
@@ -5,7 +5,8 @@ public class Main {
     public static void main(String[] args) {
 
         Scanner sc = new Scanner(System.in);
-        Product product = new Product();
+        Product product = null;
+
         int opcao = 0;
 
         while (opcao != 5) {
@@ -24,46 +25,64 @@ public class Main {
 
                 case 1:
                     System.out.print("Insira o id do produto: ");
-                    product.id = sc.nextInt();
+                    int id = sc.nextInt();
                     sc.nextLine();
 
                     System.out.print("Insira o nome do produto: ");
-                    product.name = sc.nextLine();
+                    String name = sc.nextLine();
 
                     System.out.print("Insira o preço do produto: ");
-                    product.price = sc.nextDouble();
+                    double price = sc.nextDouble();
 
                     System.out.print("Insira quantidade em estoque: ");
-                    product.quantity = sc.nextInt();
+                    int quantity = sc.nextInt();
+
+                    product = new Product(id, name, price, quantity);
 
                     System.out.println("Produto cadastrado com sucesso!");
                     break;
 
                 case 2:
-                    product.showData();
+                    if (product != null) {
+                        product.showData();
+                    } else {
+                        System.out.println("Nenhum produto cadastrado.");
+                    }
                     break;
 
                 case 3:
-                    System.out.print("Novo id: ");
-                    int newId = sc.nextInt();
-                    sc.nextLine();
+                    if (product != null) {
 
-                    System.out.print("Novo nome: ");
-                    String newName = sc.nextLine();
+                        System.out.print("Novo id: ");
+                        int newId = sc.nextInt();
+                        sc.nextLine();
 
-                    System.out.print("Novo preço: ");
-                    double newPrice = sc.nextDouble();
+                        System.out.print("Novo nome: ");
+                        String newName = sc.nextLine();
 
-                    System.out.print("Nova quantidade: ");
-                    int newQuantity = sc.nextInt();
+                        System.out.print("Novo preço: ");
+                        double newPrice = sc.nextDouble();
 
-                    product.updateData(newId, newName, newPrice, newQuantity);
-                    System.out.println("Produto atualizado com sucesso!");
+                        System.out.print("Nova quantidade: ");
+                        int newQuantity = sc.nextInt();
+
+                        product.updateData(newId, newName, newPrice, newQuantity);
+
+                        System.out.println("Produto atualizado com sucesso!");
+
+                    } else {
+                        System.out.println("Nenhum produto cadastrado.");
+                    }
                     break;
 
                 case 4:
-                    product.deleteData();
-                    System.out.println("Produto excluído com sucesso!");
+                    if (product != null) {
+                        product.deleteData();
+                        product = null;
+                        System.out.println("Produto excluído com sucesso!");
+                    } else {
+                        System.out.println("Nenhum produto cadastrado.");
+                    }
                     break;
 
                 case 5:
